Refactor Incremental helper
Consolidate `IncrementalCompile` and `Incremental`
